requirements.txt
Django~=3.0
django-background-tasks
requests
twilio
scikit-learn
pandas
django-chartjs